feat(ai): resolve AI TODOs
