+++ src/Crowdin.Api/AI/AiReportSchemaBase.cs
@@ -1,0 +1,11 @@
+
+using JetBrains.Annotations;
+
+namespace Crowdin.Api.AI
+{
+    [PublicAPI]
+    public class AiReportSchemaBase
+    {
+        
+    }
+}
