+++ src/Crowdin.Api/AI/AiReportType.cs
@@ -1,0 +1,13 @@
+
+using System.ComponentModel;
+using JetBrains.Annotations;
+
+namespace Crowdin.Api.AI
+{
+    [PublicAPI]
+    public enum AiReportType
+    {
+        [Description("tokens-usage-raw-data")]
+        TokensUsageRawData
+    }
+}
